backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF processing
PyPDF2==3.0.1

# Embeddings
sentence-transformers==2.2.2
openai==1.3.5

# Vector stores
chromadb==0.4.15
faiss-cpu>=1.7.4.post2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
requests==2.31.0

# LangChain (optional, for advanced features)
langchain==0.0.350
langchain-openai==0.0.2

# Google Generative AI (for Gemini)
google-generativeai>=0.3.0

